Testing not fixed LOL
Testing still not working as created objects are not available. However many logic errors fixed in the cart and order controllers to ensure that cart contents are checked before attempting to remove items, and that orders are only submitted when a cart actually contains items. ToString methods added to cart and item classes to inspect contents in real time for testing purposes.